// npm/ng-packs/packages/cms-kit/admin/config/src/enums/policy-names.ts
export enum eCmsKitAdminPolicyNames {
  Cms = 'CmsKit.Comments || CmsKit.Tags || CmsKit.Pages || CmsKit.Blogs || CmsKit.BlogPosts || CmsKit.Menus || CmsKit.GlobalResources',
  Comments = 'CmsKit.Comments',
  Tags = 'CmsKit.Tags',
  Pages = 'CmsKit.Pages',
  Blogs = 'CmsKit.Blogs',
  BlogPosts = 'CmsKit.BlogPosts',
  Menus = 'CmsKit.Menus',
  GlobalResources = 'CmsKit.GlobalResources',
}