// npm/ng-packs/packages/cms-kit/admin/config/src/enums/route-names.ts
export enum eCmsKitAdminRouteNames {
  Cms = 'CmsKit::Menu:CMS',
  Comments = 'CmsKit::CmsKit.Comments',
  Tags = 'CmsKit::CmsKit.Tags',
  Pages = 'CmsKit::Pages',
  Blogs = 'CmsKit::Blogs',
  BlogPosts = 'CmsKit::BlogPosts',
  Menus = 'CmsKit::Menus',
  GlobalResources = 'CmsKit::GlobalResources',
}